More fixes with sandstorm

+++ src/main/java/com/Krevik/Networking/PacketSandstormUpdatedOnClient.java
@@ -15,14 +15,14 @@ import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 public class PacketSandstormUpdatedOnClient implements IMessage {
 
     private boolean isSandstorm;
-    private float mX;
+    private double mX;
     private float sandstormTime;
-    private float mZ;
+    private double mZ;
 
     public PacketSandstormUpdatedOnClient() {
     }
 
-    public PacketSandstormUpdatedOnClient(boolean isSandstorm1, float mx,float sandstormtime,float mz) {
+    public PacketSandstormUpdatedOnClient(boolean isSandstorm1, double mx,float sandstormtime,double mz) {
         isSandstorm=isSandstorm1;
         this.mX=mx;
         this.sandstormTime=sandstormtime;
@@ -32,17 +32,17 @@ public class PacketSandstormUpdatedOnClient implements IMessage {
     @Override
     public void fromBytes(ByteBuf buf) {
         isSandstorm=buf.readBoolean();
-        mX=buf.readFloat();
+        mX=buf.readDouble();
         sandstormTime=buf.readFloat();
-        mZ=buf.readFloat();
+        mZ=buf.readDouble();
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
         buf.writeBoolean(isSandstorm);
-        buf.writeFloat(mX);
+        buf.writeDouble(mX);
         buf.writeFloat(sandstormTime);
-        buf.writeFloat(mZ);
+        buf.writeDouble(mZ);
     }
 
     public static class Handler implements IMessageHandler<PacketSandstormUpdatedOnClient, IMessage> {
